Error reporting on imports, exports, and ZIP

+++ Gladys/Models/Model/Model+ImportExport.swift
@@ -10,60 +10,62 @@ import Foundation
 import ZIPFoundation
 
 extension Model {
-	
-	static func importData(from url: URL, completion: @escaping (Bool)->Void) {
-		NSLog("URL for importing: \(url.path)")
 
+	private static func bringInItem(_ item: ArchivedDropItem, from url: URL, using fm: FileManager, moveItem: Bool) throws -> Bool {
+
+		let remotePath = url.appendingPathComponent(item.uuid.uuidString)
+		if !fm.fileExists(atPath: remotePath.path) {
+			log("Warning: Item \(item.uuid) declared but not found on imported archive, skipped")
+			return false
+		}
+
+		let localPath = item.folderUrl
+		if fm.fileExists(atPath: localPath.path) {
+			try fm.removeItem(at: localPath)
+		}
+
+		if moveItem {
+			try fm.moveItem(at: remotePath, to: localPath)
+		} else {
+			try fm.copyItem(at: remotePath, to: localPath)
+		}
+
+		item.needsReIngest = true
+		item.markUpdated()
+		item.cloudKitRecord = nil
+		for typeItem in item.typeItems {
+			typeItem.cloudKitRecord = nil
+		}
+
+		return true
+	}
+
+	static func importData(from url: URL, removingOriginal: Bool) throws {
 		let fm = FileManager.default
 		defer {
-			try? fm.removeItem(at: url)
+			if removingOriginal {
+				try? fm.removeItem(at: url)
+			}
+			save()
+			NotificationCenter.default.post(name: .ExternalDataUpdated, object: nil)
 		}
 
-		guard
-			let data = try? Data(contentsOf: url.appendingPathComponent("items.json"), options: [.alwaysMapped]),
-			let itemsInPackage = try? JSONDecoder().decode(Array<ArchivedDropItem>.self, from: data)
-			else {
-				completion(false)
-				return
-		}
-
-		var itemsImported = 0
+		let data = try Data(contentsOf: url.appendingPathComponent("items.json"), options: [.alwaysMapped])
+		let itemsInPackage = try JSONDecoder().decode(Array<ArchivedDropItem>.self, from: data)
 
 		for item in itemsInPackage.reversed() {
-
 			if let i = drops.index(of: item) {
 				if drops[i].updatedAt >= item.updatedAt {
 					continue
 				}
-				drops[i] = item
+				if try bringInItem(item, from: url, using: fm, moveItem: removingOriginal) {
+					drops[i] = item
+				}
 			} else {
-				drops.insert(item, at: 0)
+				if try bringInItem(item, from: url, using: fm, moveItem: removingOriginal) {
+					drops.insert(item, at: 0)
+				}
 			}
-
-			itemsImported += 1
-			item.needsReIngest = true
-			item.markUpdated()
-
-			let localPath = item.folderUrl
-			if fm.fileExists(atPath: localPath.path) {
-				try! fm.removeItem(at: localPath)
-			}
-
-			let remotePath = url.appendingPathComponent(item.uuid.uuidString)
-			try! fm.moveItem(at: remotePath, to: localPath)
-
-			item.cloudKitRecord = nil
-			for typeItem in item.typeItems {
-				typeItem.cloudKitRecord = nil
-			}
-		}
-
-		DispatchQueue.main.async {
-			if itemsImported > 0 {
-				save()
-				NotificationCenter.default.post(name: .ExternalDataUpdated, object: nil)
-			}
-			completion(true)
 		}
 	}
 
@@ -79,89 +81,108 @@ extension Model {
 		}
 	}
 
+	////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
 	@discardableResult
-	static func createArchive(completion: @escaping (URL)->Void) -> Progress {
+	static func createArchive(completion: @escaping (URL?, Error?) -> Void) -> Progress {
 		let eligibleItems = eligibleDropsForExport
 		let count = 2 + eligibleItems.count
 		let p = Progress(totalUnitCount: Int64(count))
 
 		DispatchQueue.global(qos: .userInitiated).async {
-
-			let fm = FileManager.default
-			let tempPath = Model.temporaryDirectoryUrl.appendingPathComponent("Gladys Archive.gladysArchive")
-			if fm.fileExists(atPath: tempPath.path) {
-				try! fm.removeItem(at: tempPath)
+			do {
+				try createArchiveThread(progress: p, eligibleItems: eligibleItems, completion: completion)
+			} catch {
+				completion(nil, error)
 			}
-
-			var delegate: FileManagerFilter? = FileManagerFilter()
-			fm.delegate = delegate
-
-			p.completedUnitCount += 1
-
-			try! fm.createDirectory(at: tempPath, withIntermediateDirectories: true, attributes: nil)
-			for item in eligibleItems {
-				let uuidString = item.uuid.uuidString
-				let sourceForItem = Model.appStorageUrl.appendingPathComponent(uuidString)
-				let destinationForItem = tempPath.appendingPathComponent(uuidString)
-				try! fm.copyItem(at: sourceForItem, to: destinationForItem)
-				p.completedUnitCount += 1
-			}
-
-			fm.delegate = nil
-			delegate = nil
-
-			let data = try! JSONEncoder().encode(eligibleItems)
-			try! data.write(to: tempPath.appendingPathComponent("items.json"))
-			p.completedUnitCount += 1
-
-			completion(tempPath)
 		}
 
 		return p
 	}
 
+	static private func createArchiveThread(progress p: Progress, eligibleItems: [ArchivedDropItem], completion: @escaping (URL?, Error?) -> Void) throws {
+		let fm = FileManager.default
+		let tempPath = Model.temporaryDirectoryUrl.appendingPathComponent("Gladys Archive.gladysArchive")
+		if fm.fileExists(atPath: tempPath.path) {
+			try fm.removeItem(at: tempPath)
+		}
+
+		var delegate: FileManagerFilter? = FileManagerFilter()
+		fm.delegate = delegate
+
+		p.completedUnitCount += 1
+
+		try fm.createDirectory(at: tempPath, withIntermediateDirectories: true, attributes: nil)
+		for item in eligibleItems {
+			let uuidString = item.uuid.uuidString
+			let sourceForItem = Model.appStorageUrl.appendingPathComponent(uuidString)
+			let destinationForItem = tempPath.appendingPathComponent(uuidString)
+			try fm.copyItem(at: sourceForItem, to: destinationForItem)
+			p.completedUnitCount += 1
+		}
+
+		fm.delegate = nil
+		delegate = nil
+
+		let data = try JSONEncoder().encode(eligibleItems)
+		try data.write(to: tempPath.appendingPathComponent("items.json"))
+		p.completedUnitCount += 1
+
+		completion(tempPath, nil)
+	}
+
+	////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
 	@discardableResult
-	static func createZip(completion: @escaping (URL)->Void) -> Progress {
+	static func createZip(completion: @escaping (URL?, Error?)->Void) -> Progress {
 
 		let dropsCopy = eligibleDropsForExport
 		let itemCount = Int64(1 + dropsCopy.count)
 		let p = Progress(totalUnitCount: itemCount)
 
-		let tempPath = Model.temporaryDirectoryUrl.appendingPathComponent("Gladys.zip")
-
 		DispatchQueue.global(qos: .userInitiated).async {
-
-			let fm = FileManager.default
-			if fm.fileExists(atPath: tempPath.path) {
-				try! fm.removeItem(at: tempPath)
+			do {
+				try createZipThread(dropsCopy: dropsCopy, progress: p, completion: completion)
+			} catch {
+				completion(nil, error)
 			}
-
-			p.completedUnitCount += 1
-
-			if let archive = Archive(url: tempPath, accessMode: .create) {
-				for item in dropsCopy {
-					let dir = item.displayTitleOrUuid.filenameSafe
-
-					if item.typeItems.count == 1 {
-						let typeItem = item.typeItems.first!
-						self.addZipItem(typeItem, directory: nil, name: dir, in: archive)
-
-					} else {
-						for typeItem in item.typeItems {
-							self.addZipItem(typeItem, directory: dir, name: typeItem.typeDescription, in: archive)
-						}
-					}
-					p.completedUnitCount += 1
-				}
-			}
-
-			completion(tempPath)
 		}
 
 		return p
 	}
 
-	static private func addZipItem(_ typeItem: ArchivedDropItemType, directory: String?, name: String, in archive: Archive) {
+	static func createZipThread(dropsCopy: [ArchivedDropItem], progress p: Progress, completion: @escaping (URL?, Error?)->Void) throws {
+
+		let tempPath = Model.temporaryDirectoryUrl.appendingPathComponent("Gladys.zip")
+
+		let fm = FileManager.default
+		if fm.fileExists(atPath: tempPath.path) {
+			try fm.removeItem(at: tempPath)
+		}
+
+		p.completedUnitCount += 1
+
+		if let archive = Archive(url: tempPath, accessMode: .create) {
+			for item in dropsCopy {
+				let dir = item.displayTitleOrUuid.filenameSafe
+
+				if item.typeItems.count == 1 {
+					let typeItem = item.typeItems.first!
+					try addZipItem(typeItem, directory: nil, name: dir, in: archive)
+
+				} else {
+					for typeItem in item.typeItems {
+						try addZipItem(typeItem, directory: dir, name: typeItem.typeDescription, in: archive)
+					}
+				}
+				p.completedUnitCount += 1
+			}
+		}
+
+		completion(tempPath, nil)
+	}
+
+	static private func addZipItem(_ typeItem: ArchivedDropItemType, directory: String?, name: String, in archive: Archive) throws {
 
 		var bytes: Data?
 		if typeItem.isWebURL, let url = typeItem.encodedUrl, let data = url.urlFileContent {
@@ -172,7 +193,7 @@ extension Model {
 		}
 		if let B = bytes ?? typeItem.bytes {
 			let timmedName = typeItem.prepareFilename(name: name, directory: directory)
-			try? archive.addEntry(with: timmedName, type: .file, uncompressedSize: UInt32(B.count)) { pos, size -> Data in
+			try archive.addEntry(with: timmedName, type: .file, uncompressedSize: UInt32(B.count)) { pos, size -> Data in
 				return B[pos ..< pos+size]
 			}
 		}
